🔥 Remove README from documentation
There is no way that Sphinx and the root README can use the same images cause both have a different root.
In ./README.rst you will use `./docs/_static...`
This will work for the repo readme, but Sphinx will shout in your face, that it can't find the image.
Than you can do a reference with `../docs/_static...`
Now Sphinx is happy, but the links in the README are broken -.-