minor (test only) renaming: "valid" != "well-formed"
I'd like to introduce a little differentiation with a nomenclature very similar to XML/HTML.

- we call those expressions **"well-formed"** which obey the general syntax rules (not well-formed expressions raise an SyntaxError)
- we call those expressions **"valid"** (or at least _not_ invalid) that are both well-formed and can have a meaningful evaluation outcome

Motivation: ahbicht shall be used to flag well-formed expressions that are invalid though.

E.g.
> `Muss ([77] u[78]) u (([61]u [584]) o[583])"`

is well-formed (obeys the syntax rules, has balanced brackets and uses only allowed operators and tokens) but still the expression is invalid, because the part `([61]u [584]) o[583]` cannot be evaluated.
https://github.com/Hochfrequenz/edifact-templates/issues/133

This raises errors at evaluation runtime which I'd like to prevent and spot in advance.

Same