chore: restrict marshmallow to <v4 (#580)
* chore: restrict marshmallow to <v4

because a lots of stuff goes wrong in v4: see #575 
probably we should get rid of it together with attrs and just use pydantic where we'd have none of the problems (e.g. also not #218 and our dependency on https://github.com/fuhrysteve/marshmallow-jsonschema/issues/169 which we really should remove in the long term.

* reset requirements for doc creation

---------

Co-authored-by: DeltaDaniel <139119540+DeltaDaniel@users.noreply.github.com>